requirements.txt
pandas==2.3.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
chardet==5.2.0